+++ libft/ft_itoa_unsigned.c
@@ -6,13 +6,13 @@
 /*   By: gandrade <gandrade@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/02 22:25:16 by gandrade          #+#    #+#             */
-/*   Updated: 2021/07/26 18:09:14 by gandrade         ###   ########.fr       */
+/*   Updated: 2021/07/26 22:58:32 by gandrade         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static	void	ft_nbrstr(char *str, unsigned int n, size_t size)
+static	void	ft_nbrstr(char *str, unsigned int n, int size)
 {
 	str[size] = '\0';
 	while (size--)
@@ -25,7 +25,7 @@ static	void	ft_nbrstr(char *str, unsigned int n, size_t size)
 char	*ft_itoa_unsigned(unsigned int n)
 {
 	char			*res;
-	size_t			size;
+	int				size;
 	unsigned int	aux;
 
 	size = 1;
